SSH module version 2.0
Simplify the existing module by removing all code that deals with setting specific versions and move to in-module hiera data. This allows for full support of all `sshd_config`
options without the need for special casing. We now only make the settings changes that the users instructs rather then picking a few safe options from `params.pp` (which is
also gone anyway). This means users are only managing the settings they want to manage and avoid accidentally clobbering new settings from the vendor which may contain
important security fixes. All settings for the config file are now passed in the parameter `settings`. This replaces the `extra_config` parameter and all of the individual
settings parameters that used to exist. Since this is a breaking change the version number has been bumped and users of the previous module need to either change their
profile/hiera usage or stay on the old version.